🚚 Move the SearchModule component in the correct folder (landing page -> dashboard)
